+++ src/libfrozen/machines/io/tcp/tcp.c
@@ -81,7 +81,7 @@ static ssize_t tcp_control_start(machine_t *machine){ // {{{
 	return 0;
 } // }}}
 
-static int tcp_init(machine_t *machine){ // {{{
+static ssize_t tcp_init(machine_t *machine){ // {{{
 	tcp_userdata        *userdata;
 	
 	if(tcp_global_inited == 0){
@@ -99,7 +99,7 @@ static int tcp_init(machine_t *machine){ // {{{
 	userdata->tcp_addr = strdup("127.0.0.1");
 	return 0;
 } // }}}
-static int tcp_destroy(machine_t *machine){ // {{{
+static ssize_t tcp_destroy(machine_t *machine){ // {{{
 	tcp_userdata          *userdata          = (tcp_userdata *)machine->userdata;
 	
 	if(userdata->tcp_addr)
@@ -108,7 +108,7 @@ static int tcp_destroy(machine_t *machine){ // {{{
 	free(userdata);
 	return 0;
 } // }}}
-static int tcp_configure(machine_t *machine, hash_t *config){ // {{{
+static ssize_t tcp_configure(machine_t *machine, hash_t *config){ // {{{
 	ssize_t                ret;
 	tcp_userdata          *userdata          = (tcp_userdata *)machine->userdata;
 	
@@ -153,7 +153,7 @@ machine_t tcp_proto = {
 	},
 };
 
-static int tcp_child_init(machine_t *machine){ // {{{
+static ssize_t tcp_child_init(machine_t *machine){ // {{{
 	int                   *socket;
 	tcp_child_userdata    *userdata;
 	
@@ -173,7 +173,7 @@ static int tcp_child_init(machine_t *machine){ // {{{
 	userdata->socket = *socket;
 	return 0;
 } // }}}
-static int tcp_child_destroy(machine_t *machine){ // {{{
+static ssize_t tcp_child_destroy(machine_t *machine){ // {{{
 	tcp_child_userdata    *userdata          = (tcp_child_userdata *)machine->userdata;
 	
 	if(userdata->primary == 1){
